Enhance 3D immersed-boundary method: add geometry and stencil functions, refactor multi-direct forcing, and remove obsolete 3D module files.
