--- vivsim/ib3d/__init__.py
@@ -1,28 +1,0 @@
-"""3D immersed-boundary package exports."""
-
-from .geometry import (
-    get_area,
-    get_marker_dA,
-    get_surface_area,
-    get_triangle_areas,
-    get_vertex_dA,
-)
-from .kernels import kernel_cosine_4pt, kernel_peskin_3pt, kernel_peskin_4pt
-from .mdf import multi_direct_forcing
-from .stencil import get_ib_stencil, interpolate, spread
-
-
-__all__ = [
-    "get_area",
-    "get_marker_dA",
-    "get_surface_area",
-    "get_triangle_areas",
-    "get_vertex_dA",
-    "kernel_cosine_4pt",
-    "kernel_peskin_3pt",
-    "kernel_peskin_4pt",
-    "multi_direct_forcing",
-    "get_ib_stencil",
-    "interpolate",
-    "spread",
-]
